Fix warning: implicit conversion changes signedness: 'const int' to 'size_type' (aka 'unsigned long') [-Wsign-conversion]
visp/modules/core/include/visp3/core/vpImageFilter.h:2609:35: warning: implicit conversion changes signedness: 'const int' to 'size_type' (aka 'unsigned long') [-Wsign-conversion]
 2609 |     std::vector<OutputType> Idiff(size);
      |                             ~~~~~ ^~~~
visp/modules/core/include/visp3/core/vpImageFilter.h:2087:7: note: in instantiation of function template specialization 'vpImageFilter::gradientFilterXMonothread<unsigned char, true, double>' requested here
 2087 |       gradientFilterXMonothread(I, GIx, filter, p_mask);
      |       ^
visp/modules/core/include/visp3/core/vpImageFilter.h:2429:5: note: in instantiation of function template specialization 'vpImageFilter::gradientFilterX<unsigned char, double, true>' requested here
 2429 |     gradientFilterX(I, GIx, nbThread, p_mask, type);
      |     ^
visp/modules/core/test/image/testHSVGradient.cpp:304:26: note: in instantiation of function template specialization 'vpImageFilter::gradientFilter<unsigned char, double, true>' requested here
  304 |           vpImageFilter::gradientFilter(input.second.m_I, GIx, GIy, nbThread, p_mask, type);
      |                          ^
In file included from visp/modules/core/test/image/testHSVGradient.cpp:45:
visp/modules/core/include/visp3/core/vpImageFilter.h:2620:22: warning: implicit conversion changes signedness: 'int' to 'size_type' (aka 'unsigned long') [-Wsign-conversion]
 2620 |           Idiff[iter + offsetIdiff] = futureDiff;
      |           ~~~~~ ~~~~~^~~~~~~~~~~~~
visp/modules/core/include/visp3/core/vpImageFilter.h:2630:73: warning: implicit conversion changes signedness: 'int' to 'size_type' (aka 'unsigned long') [-Wsign-conversion]
 2630 |               gradient += filter[i + 1] * (Idiff[offset] + Idiff[offset - 1]);
      |                                                            ~~~~~ ~~~~~~~^~~
visp/modules/core/include/visp3/core/vpImageFilter.h:2630:50: warning: implicit conversion changes signedness: 'int' to 'size_type' (aka 'unsigned long') [-Wsign-conversion]
 2630 |               gradient += filter[i + 1] * (Idiff[offset] + Idiff[offset - 1]);
      |                                            ~~~~~ ^~~~~~
visp/modules/core/include/visp3/core/vpImageFilter.h:2630:36: warning: implicit conversion changes signedness: 'int' to 'size_type' (aka 'unsigned long') [-Wsign-conversion]
 2630 |               gradient += filter[i + 1] * (Idiff[offset] + Idiff[offset - 1]);
      |                           ~~~~~~ ~~^~~
visp/modules/core/include/visp3/core/vpImageFilter.h:2750:26: warning: implicit conversion changes signedness: 'unsigned int' to 'const int' [-Wsign-conversion]
 2750 |     const int nbRows = I.getRows(), nbCols = I.getCols();
      |               ~~~~~~   ~~^~~~~~~~~
visp/modules/core/include/visp3/core/vpImageFilter.h:2090:7: note: in instantiation of function template specialization 'vpImageFilter::gradientFilterXMultithread<unsigned char, true, double>' requested here
 2090 |       gradientFilterXMultithread(I, GIx, filter, nbThread, p_mask);
      |       ^
visp/modules/core/include/visp3/core/vpImageFilter.h:2429:5: note: in instantiation of function template specialization 'vpImageFilter::gradientFilterX<unsigned char, double, true>' requested here
 2429 |     gradientFilterX(I, GIx, nbThread, p_mask, type);
      |     ^

+++ modules/core/include/visp3/core/vpImageFilter.h
@@ -1549,7 +1549,7 @@ public:
 #ifdef VISP_HAVE_OPENMP
       }
 #endif
-    }
+      }
   }
 
   /**
@@ -2606,7 +2606,7 @@ private:
     const int size = static_cast<int>(I.getSize());
     const int offsetIdiff = nbCols;
 
-    std::vector<OutputType> Idiff(size);
+    std::vector<OutputType> Idiff(static_cast<size_t>(size));
     initGradientFilterDifferenceImageX(I, Idiff);
     const int resetCounter = nbCols - 1;
     const int stopIter = size - (nbCols + 1);
@@ -2616,8 +2616,8 @@ private:
         // Computing the amplitude of the difference
         OutputType futureDiff = 0.;
         if (checkBooleanPatch(p_mask, iter + offsetIdiff, idCol, nbRows, nbCols, true)) {
-          futureDiff = static_cast<OutputType>(I.bitmap[iter + nbCols +1].V - I.bitmap[iter + offsetIdiff].V);
-          Idiff[iter + offsetIdiff] = futureDiff;
+          futureDiff = static_cast<OutputType>(I.bitmap[static_cast<size_t>(iter + nbCols +1)].V - I.bitmap[static_cast<size_t>(iter + offsetIdiff)].V);
+          Idiff[static_cast<size_t>(iter + offsetIdiff)] = futureDiff;
         }
       }
       if (counter) {
@@ -2627,7 +2627,7 @@ private:
             int offset = iter - nbCols; // Looking in the row above first
             for (int i = -1; i <= 1; ++i) {
               // Kind of  +/- (I[r + i][c + 1] - I[r + i][c]) +/- (I[r + i][c] - I[r + i][c - 1])
-              gradient += filter[i + 1] * (Idiff[offset] + Idiff[offset - 1]);
+              gradient += filter[static_cast<size_t>(i + 1)] * (Idiff[static_cast<size_t>(offset)] + Idiff[static_cast<size_t>(offset - 1)]);
               offset += nbCols; // Preparing to look in the next row
             }
             GI.bitmap[iter] = gradient;
@@ -2670,11 +2670,11 @@ private:
     const vpImage<bool> *p_mask = nullptr
   )
   {
-    const int nbRows = I.getRows(), nbCols = I.getCols();
-    const int size = I.getSize();
+    const int nbRows = static_cast<int>(I.getRows()), nbCols = static_cast<int>(I.getCols());
+    const int size = static_cast<int>(I.getSize());
     const int offsetIdiff = 1;
 
-    std::vector<OutputType> Idiff(size);
+    std::vector<OutputType> Idiff(static_cast<size_t>(size));
     initGradientFilterDifferenceImageY(I, Idiff);
     const int resetCounter = nbCols - 1;
     const int stopIter = size - (nbCols + 1);
@@ -2684,8 +2684,8 @@ private:
       OutputType futureDiff = 0.;
 
       if (checkBooleanPatch(p_mask, iter + offsetIdiff, iterSign, nbRows, nbCols, false)) {
-        futureDiff = static_cast<OutputType>(I.bitmap[iter + nbCols +1].V - I.bitmap[iter + offsetIdiff].V);
-        Idiff[iter + offsetIdiff] = futureDiff;
+        futureDiff = static_cast<OutputType>(I.bitmap[static_cast<size_t>(iter + nbCols +1)].V - I.bitmap[static_cast<size_t>(iter + offsetIdiff)].V);
+        Idiff[static_cast<size_t>(iter + offsetIdiff)] = futureDiff;
       }
 
       if (counter) {
@@ -2694,7 +2694,7 @@ private:
             OutputType gradient = 0.;
             for (int i = -1; i <= 1; ++i) {
               // Kind of +/- (I[r + 1][c + i] - I[r][c + 1]) +/- (I[r][c + i] - I[r - 1][c + 1])
-              gradient += filter[i + 1] * (Idiff[iter + i] + Idiff[iter - nbCols + i]);
+              gradient += filter[static_cast<size_t>(i + 1)] * (Idiff[static_cast<size_t>(iter + i)] + Idiff[static_cast<size_t>(iter - nbCols + i)]);
             }
             GI.bitmap[iter] = gradient;
           }
@@ -2720,25 +2720,25 @@ private:
     const int &istart, const int &iam
   )
   {
-    const int nbCols = I.getCols();
+    const int nbCols = static_cast<int>(I.getCols());
 
     if (iam > 0) {
       Idiff[0] = static_cast<OutputType>(I.bitmap[istart - nbCols + 1].V - I.bitmap[istart - nbCols].V);
     }
 
     // Computing the difference and sign for row 1 column 0, which corresponds to the current row of the image
-    Idiff[nbCols] = static_cast<OutputType>(I.bitmap[istart + 1].V - I.bitmap[istart].V);
+    Idiff[static_cast<size_t>(nbCols)] = static_cast<OutputType>(I.bitmap[istart + 1].V - I.bitmap[istart].V);
 
     for (int iter = 1; iter < nbCols - 1; ++iter) {
       if (iam > 0) {
         // Computing the difference and sign for row 0, which corresponds to the previous row of the image
         OutputType distanceRow0 = static_cast<OutputType>(I.bitmap[istart - nbCols + iter + 1].V - I.bitmap[istart - nbCols + iter].V);
-        Idiff[iter] = distanceRow0;
+        Idiff[static_cast<size_t>(iter)] = distanceRow0;
       }
 
       // Computing the difference and sign for row 1, which corresponds to the current row of the image
       OutputType distanceRow1 = static_cast<OutputType>(I.bitmap[istart + iter + 1].V - I.bitmap[istart + iter].V);
-      Idiff[nbCols + iter] = distanceRow1;
+      Idiff[static_cast<size_t>(nbCols + iter)] = distanceRow1;
     }
   }
 
@@ -2747,7 +2747,7 @@ private:
     const vpImage<vpHSV<HSVType, useFullScale>> &I, vpImage<OutputType> &GI, const std::vector<OutputType> &filter,
     const int &maxNbThread, const vpImage<bool> *p_mask = nullptr)
   {
-    const int nbRows = I.getRows(), nbCols = I.getCols();
+    const int nbRows = static_cast<int>(I.getRows()), nbCols = static_cast<int>(I.getCols());
     const int offsetIdiff = nbCols;
     const int resetCounter = nbCols - 1;
     const int nrows(nbRows - 1);
@@ -2777,8 +2777,8 @@ private:
       }
       istop = istart + irows * nbCols;
 
-      std::vector<OutputType> Idiff((irows + 2) * nbCols);
-      std::vector<OutputType> GItemp(irows * nbCols);
+      std::vector<OutputType> Idiff(static_cast<size_t>((irows + 2) * nbCols));
+      std::vector<OutputType> GItemp(static_cast<size_t>(irows * nbCols));
       initGradientFilterDifferenceImageX(I, Idiff, istart, iam);
 
       int counter = resetCounter, idCol = 0;
@@ -2788,8 +2788,8 @@ private:
           // Computing the amplitude of the difference
           OutputType futureDiff = 0.;
           if (checkBooleanPatch(p_mask, iter + offsetIdiff, idCol, nbRows, nbCols, true)) {
-            futureDiff = static_cast<OutputType>(I.bitmap[iter + nbCols +1].V - I.bitmap[iter + offsetIdiff].V);
-            Idiff[iter + 2 * offsetIdiff - istart] = futureDiff;
+            futureDiff = static_cast<OutputType>(I.bitmap[static_cast<size_t>(iter + nbCols +1)].V - I.bitmap[static_cast<size_t>(iter + offsetIdiff)].V);
+            Idiff[static_cast<size_t>(iter + 2 * offsetIdiff - istart)] = futureDiff;
           }
         }
         if (counter) {
@@ -2799,10 +2799,10 @@ private:
               int offset = iter - istart; // Looking in the row above first
               for (int i = -1; i <= 1; ++i) {
                 // Kind of  +/- (I[r + i][c + 1] - I[r + i][c]) +/- (I[r + i][c] - I[r + i][c - 1])
-                gradient += filter[i + 1] * (Idiff[offset] + Idiff[offset - 1]);
+                gradient += filter[static_cast<size_t>(i + 1)] * (Idiff[static_cast<size_t>(offset)] + Idiff[static_cast<size_t>(offset - 1)]);
                 offset += nbCols; // Preparing to look in the next row
               }
-              GItemp[iter - istart] = gradient;
+              GItemp[static_cast<size_t>(iter - istart)] = gradient;
             }
           }
           --counter;
@@ -2831,17 +2831,17 @@ private:
     const int &istart
   )
   {
-    const int nbCols = I.getCols();
+    const int nbCols = static_cast<int>(I.getCols());
     // Computing the sign and distance for the first row, which corresponds to the row above the beginning of the gradient computation in the thread
     int idDiff = 0;
     for (int iter = istart - nbCols; iter < istart; ++iter) {
       OutputType distance = static_cast<OutputType>(I.bitmap[iter + nbCols].V - I.bitmap[iter].V);
-      Idiff[idDiff] = distance;
+      Idiff[static_cast<size_t>(idDiff)] = distance;
       ++idDiff;
     }
     // Computing the distance and sign for I[1][0]
-    OutputType distance = static_cast<OutputType>(I.bitmap[nbCols + nbCols].V - I.bitmap[nbCols].V);
-    Idiff[nbCols] = distance;
+    OutputType distance = static_cast<OutputType>(I.bitmap[static_cast<size_t>(nbCols + nbCols)].V - I.bitmap[static_cast<size_t>(nbCols)].V);
+    Idiff[static_cast<size_t>(nbCols)] = distance;
   }
 
   template <typename HSVType, bool useFullScale, typename OutputType>
@@ -2849,7 +2849,7 @@ private:
     const vpImage<vpHSV<HSVType, useFullScale>> &I, vpImage<OutputType> &GI, const std::vector<OutputType> &filter,
     const int &maxNbThread, const vpImage<bool> *p_mask = nullptr)
   {
-    const int nbRows = I.getRows(), nbCols = I.getCols();
+    const int nbRows = static_cast<int>(I.getRows()), nbCols = static_cast<int>(I.getCols());
     const int offsetIdiff = 1;
     const int resetCounter = nbCols - 1;
     const int nrows(nbRows - 1);
@@ -2879,8 +2879,8 @@ private:
       }
       istop = istart + irows * nbCols;
 
-      std::vector<OutputType> Idiff((irows + 2) * nbCols);
-      std::vector<OutputType> GItemp(irows * nbCols);
+      std::vector<OutputType> Idiff(static_cast<size_t>((irows + 2) * nbCols));
+      std::vector<OutputType> GItemp(static_cast<size_t>(irows * nbCols));
 
       if (iam == 0) {
         initGradientFilterDifferenceImageY(I, Idiff);
@@ -2896,8 +2896,8 @@ private:
         OutputType futureDiff = 0.;
 
         if (checkBooleanPatch(p_mask, iter + offsetIdiff, iterSign, nbRows, nbCols, false)) {
-          futureDiff = static_cast<OutputType>(I.bitmap[iter + nbCols +1].V - I.bitmap[iter + offsetIdiff].V);
-          Idiff[iter - istart + nbCols + offsetIdiff] = futureDiff;
+          futureDiff = static_cast<OutputType>(I.bitmap[static_cast<size_t>(iter + nbCols +1)].V - I.bitmap[static_cast<size_t>(iter + offsetIdiff)].V);
+          Idiff[static_cast<size_t>(iter - istart + nbCols + offsetIdiff)] = futureDiff;
         }
 
         if (counter) {
@@ -2906,9 +2906,9 @@ private:
               OutputType gradient = 0.;
               for (int i = -1; i <= 1; ++i) {
                 // Kind of +/- (I[r + 1][c + i] - I[r][c + 1]) +/- (I[r][c + i] - I[r - 1][c + 1])
-                gradient += filter[i + 1] * (Idiff[iter - istart + nbCols + i] + Idiff[iter - istart + i]);
+                gradient += filter[static_cast<size_t>(i + 1)] * (Idiff[static_cast<size_t>(iter - istart + nbCols + i)] + Idiff[static_cast<size_t>(iter - istart + i)]);
               }
-              GItemp[iter - istart] = gradient;
+              GItemp[static_cast<size_t>(iter - istart)] = gradient;
             }
           }
           --counter;
